+++ apps/ced-portal-be/CHANGELOG.md
@@ -1,5 +1,19 @@
 # ced-portal-be
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [27672b8]
+  - @pagopa/io-core-adapter-redis@0.0.5
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [a6df4db]
+  - @pagopa/io-core-adapter-redis@0.0.4
+
 ## 0.0.6
 
 ### Patch Changes