+++ packages/io-core-adapter-redis/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pagopa/io-core-adapter-redis
 
+## 0.0.5
+
+### Patch Changes
+
+- 27672b8: fixed connection
+
+## 0.0.4
+
+### Patch Changes
+
+- a6df4db: removed minimize connections
+
 ## 0.0.3
 
 ### Patch Changes